Prevent installing tracker and its deps
It's useless in dom0, as no user data should be accessed there.
So, save 40MB, and 5 extra systemd services that are going to only waste
CPU cycles.

For that introduce new qubes-dom0-unwanted-packages metapackage. Don't
add this dependency to any actually important package, so if user really
want to install this (or other added in the future) package, it can be
done without much harm.