debian: use dynamic variables instead of patching control file
Do not patch debian/control before the build, because it results in
orig.tar.gz being different for different Debian versions, and reprepro
is very much not happy about it (the repository has a single orig.tar.gz
for all Debian versions).
Instead, use dynamic variables generated in debian/rules.
Added benefit is that the logic is now in a single place, not duplicated
in Makefile.builder and .qubesbuilder.