Attempt key import/del every 5s until 15m timeout
Installing a qubes-contrib package immediately after adding the
repository via "qubes-repo-contrib" would fail. This was because the key
had not yet been imported and dom0 was discarding the package due to it
not being properly signed.

A shorter timeout increases the likelyhood that the key is already
available by the time the user attempts to make use of the newly added
qubes-contrib repository. This also adds the same logic key removal, but
the argument there is that the system could be shut down before the
systemd-run command had a chance to remove the key.

Fixes https://github.com/QubesOS/qubes-issues/issues/10311